requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
openai==1.54.4
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==3.6.0
sendgrid-django==4.2.0
setuptools==75.5.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
